Refactors Schedule management
part 1
