+++ src/logic/lobbyScheduling/types/IScheduleFields.ts
@@ -1,0 +1,6 @@
+import { IFieldElement } from "../../discord/interfaces/IFieldElement";
+
+export interface IScheduleFields {
+  schedules: IFieldElement[];
+  lastReactionEmojiIndex: number;
+}
