+++ src/logic/lobbyScheduling/types/IScheduleTextOptions.ts
@@ -1,0 +1,8 @@
+export interface IScheduleTestOptions {
+  days: Array<string>;
+  emojis: Array<string>;
+  title: string;
+  times: Array<string>;
+  timezoneName: string;
+  coachCount: number;
+}
