+++ src/logic/lobbyScheduling/types/IWeeklyScheduleData.ts
@@ -1,0 +1,7 @@
+export interface IWeeklyScheduleData {
+  coachCount: number;
+  daysByRegion: Array<Array<number>>;
+  timesByRegion: Array<Array<string>>;
+  channelId: string;
+  type: string;
+}
